terraform will pull from my docker cloud for images, need to add more docker images up there first, have test script to run create and destroy from docker api

+++ test.py
@@ -1,3 +1,4 @@
+import argparse
 from terraform import api
 from terraform import EXAMPLE_VARS
 import os
@@ -5,11 +6,18 @@ import os
 if __name__ == '__main__':
     ansible_dir = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'ansible')
     EXAMPLE_VARS['ansible_dir'] = ansible_dir
-    
-    print("Creating...")
-    api.create(EXAMPLE_VARS)
-    print("Created")
 
-    print("Destroying...")
-    api.destroy(EXAMPLE_VARS)
-    print("Destroyed")
+    parser = argparse.ArgumentParser(description="Test out terraform api")
+    parser.add_argument('-c', action='store_true')
+    parser.add_argument('-d', action='store_true')    
+    args = parser.parse_args()
+    
+    if args.c:
+        print("Creating...")
+        api.create(EXAMPLE_VARS)
+        print("Created")
+
+    if args.d:
+        print("Destroying...")
+        api.destroy(EXAMPLE_VARS)
+        print("Destroyed")
